# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: B-View Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List and filter expenses
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: vendor
          in: query
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            type: string
        - name: minAmount
          in: query
          schema:
            type: number
        - name: maxAmount
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated expense list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedExpenses"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get a single expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Soft-delete an expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /export/pdf:
    get:
      operationId: exportPdf
      tags: [export]
      summary: Export expenses as PDF
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: PDF file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
  /export/excel:
    get:
      operationId: exportExcel
      tags: [export]
      summary: Export expenses as Excel
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Excel file download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        amount:
          type: number
        currency:
          type: string
        vendor:
          type: string
        category:
          type: string
        department:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
          enum: [Cash, BankTransfer, POS, CreditCard, Check]
        personResponsible:
          type: string
        documentUrl:
          type: string
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdById:
          type: integer
        createdByName:
          type: string
          nullable: true
      required:
        - id
        - date
        - amount
        - currency
        - vendor
        - category
        - department
        - description
        - paymentMethod
        - personResponsible
        - createdAt
        - updatedAt
        - createdById
    CreateExpenseBody:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: NGN
        vendor:
          type: string
        category:
          type: string
        department:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
          enum: [Cash, BankTransfer, POS, CreditCard, Check]
        personResponsible:
          type: string
        documentUrl:
          type: string
          nullable: true
      required:
        - date
        - amount
        - vendor
        - category
        - department
        - description
        - paymentMethod
        - personResponsible
    UpdateExpenseBody:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
        currency:
          type: string
        vendor:
          type: string
        category:
          type: string
        department:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
          enum: [Cash, BankTransfer, POS, CreditCard, Check]
        personResponsible:
          type: string
        documentUrl:
          type: string
          nullable: true
    PaginatedExpenses:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...
    ExportResponse:
      type: object
      properties:
        downloadUrl:
          type: string
        filename:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - downloadUrl
        - filename
        - expiresAt
...