# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * B-View Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExpenseBodyPaymentMethod } from './createExpenseBodyPaymentMethod';

export interface CreateExpenseBody {
  date: string;
  amount: number;
  currency?: string;
  vendor: string;
  category: string;
  department: string;
  description: string;
  paymentMethod: CreateExpenseBodyPaymentMethod;
  personResponsible: string;
  documentUrl?: string | null;
}