# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * B-View Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpensePaymentMethod } from './expensePaymentMethod';

export interface Expense {
  id: number;
  date: string;
  amount: number;
  currency: string;
  vendor: string;
  category: string;
  department: string;
  description: string;
  paymentMethod: ExpensePaymentMethod;
  personResponsible: string;
  documentUrl?: string | null;
  deletedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
  createdById: number;
  createdByName?: string | null;
}